yapf

+++ examples/reinforcement_learning/tutorial_A3C.py
@@ -76,26 +76,27 @@ GAME = 'BipedalWalker-v2'  # BipedalWalkerHardcore-v2   BipedalWalker-v2  LunarL
 LOG_DIR = './log'  # the log file
 N_WORKERS = multiprocessing.cpu_count()  # number of workers accroding to number of cores in cpu
 # N_WORKERS = 2     # manually set number of workers
-MAX_GLOBAL_EP = 8   # number of training episodes
+MAX_GLOBAL_EP = 8  # number of training episodes
 GLOBAL_NET_SCOPE = 'Global_Net'
-UPDATE_GLOBAL_ITER = 10   # update global policy after several episodes
-GAMMA = 0.99       # reward discount factor
+UPDATE_GLOBAL_ITER = 10  # update global policy after several episodes
+GAMMA = 0.99  # reward discount factor
 ENTROPY_BETA = 0.005  # factor for entropy boosted exploration
 LR_A = 0.00005  # learning rate for actor
 LR_C = 0.0001  # learning rate for critic
 GLOBAL_RUNNING_R = []
 GLOBAL_EP = 0  # will increase during training, stop training when it >= MAX_GLOBAL_EP
 
-
 ###################  Asynchronous Advantage Actor Critic (A3C)  ####################################
+
 
 class ACNet(object):
 
-    def __init__(self, scope, globalAC=None):  
+    def __init__(self, scope, globalAC=None):
         self.scope = scope
         self.save_path = './model'
 
         w_init = tf.keras.initializers.glorot_normal(seed=None)  # initializer, glorot=xavier
+
         def get_actor(input_shape):  # policy network
             with tf.name_scope(self.scope):
                 ni = tl.layers.Input(input_shape, name='in')
@@ -103,21 +104,26 @@ class ACNet(object):
                 nn = tl.layers.Dense(n_units=300, act=tf.nn.relu6, W_init=w_init, name='la2')(nn)
                 mu = tl.layers.Dense(n_units=N_A, act=tf.nn.tanh, W_init=w_init, name='mu')(nn)
                 sigma = tl.layers.Dense(n_units=N_A, act=tf.nn.softplus, W_init=w_init, name='sigma')(nn)
-            return tl.models.Model(inputs=ni, outputs=[mu, sigma], name=scope+'/Actor')
-        self.actor = get_actor( [None, N_S])
+            return tl.models.Model(inputs=ni, outputs=[mu, sigma], name=scope + '/Actor')
+
+        self.actor = get_actor([None, N_S])
         self.actor.train()  # train mode for Dropout, BatchNorm
-        def get_critic(input_shape): # we use Value-function here, but not Q-function.
+
+        def get_critic(input_shape):  # we use Value-function here, but not Q-function.
             with tf.name_scope(self.scope):
                 ni = tl.layers.Input(input_shape, name='in')
                 nn = tl.layers.Dense(n_units=500, act=tf.nn.relu6, W_init=w_init, name='lc')(ni)
                 nn = tl.layers.Dense(n_units=300, act=tf.nn.relu6, W_init=w_init, name='lc2')(nn)
                 v = tl.layers.Dense(n_units=1, W_init=w_init, name='v')(nn)
-            return tl.models.Model(inputs=ni, outputs=v, name=scope+'/Critic')
-        self.critic = get_critic( [None, N_S])
-        self.critic.train() # train mode for Dropout, BatchNorm
-    
-    @tf.function # convert numpy functions to tf.Operations in the TFgraph, return tensor
-    def update_global(self, buffer_s, buffer_a, buffer_v_target, globalAC):  # refer to the global Actor-Crtic network for updating it with samples  
+            return tl.models.Model(inputs=ni, outputs=v, name=scope + '/Critic')
+
+        self.critic = get_critic([None, N_S])
+        self.critic.train()  # train mode for Dropout, BatchNorm
+
+    @tf.function  # convert numpy functions to tf.Operations in the TFgraph, return tensor
+    def update_global(
+            self, buffer_s, buffer_a, buffer_v_target, globalAC
+    ):  # refer to the global Actor-Crtic network for updating it with samples
         ''' update the global critic '''
         with tf.GradientTape() as tape:
             self.v = self.critic(buffer_s)
@@ -127,7 +133,6 @@ class ACNet(object):
         self.c_grads = tape.gradient(self.c_loss, self.critic.trainable_weights)
         OPT_C.apply_gradients(zip(self.c_grads, globalAC.critic.trainable_weights))  # local grads applies to global net
         # del tape # Drop the reference to the tape
-
         ''' update the global actor '''
         with tf.GradientTape() as tape:
             self.mu, self.sigma = self.actor(buffer_s)
@@ -135,7 +140,7 @@ class ACNet(object):
             self.mu, self.sigma = self.mu * A_BOUND[1], self.sigma + 1e-5
 
             normal_dist = tfd.Normal(self.mu, self.sigma)  # no tf.contrib for tf2.0
-            self.a_his = buffer_a # float32
+            self.a_his = buffer_a  # float32
             log_prob = normal_dist.log_prob(self.a_his)
             exp_v = log_prob * td  # td is from the critic part, no gradients for it
             entropy = normal_dist.entropy()  # encourage exploration
@@ -144,7 +149,7 @@ class ACNet(object):
         self.a_grads = tape.gradient(self.a_loss, self.actor.trainable_weights)
         OPT_A.apply_gradients(zip(self.a_grads, globalAC.actor.trainable_weights))  # local grads applies to global net
         return self.test  # for test purpose
-    
+
     @tf.function
     def pull_global(self, globalAC):  # run by a local, pull weights from the global nets
         for l_p, g_p in zip(self.actor.trainable_weights, globalAC.actor.trainable_weights):
@@ -158,17 +163,18 @@ class ACNet(object):
 
         with tf.name_scope('wrap_a_out'):
             self.mu, self.sigma = self.mu * A_BOUND[1], self.sigma + 1e-5
-        normal_dist = tfd.Normal(self.mu, self.sigma)   # for continuous action space
+        normal_dist = tfd.Normal(self.mu, self.sigma)  # for continuous action space
         self.A = tf.clip_by_value(tf.squeeze(normal_dist.sample(1), axis=0), *A_BOUND)
         return self.A.numpy()[0]
 
-    def save_ckpt(self): # save trained weights
+    def save_ckpt(self):  # save trained weights
         tl.files.save_npz(self.actor.trainable_weights, name='model_actor.npz')
         tl.files.save_npz(self.critic.trainable_weights, name='model_critic.npz')
 
-    def load_ckpt(self): # load trained weights
+    def load_ckpt(self):  # load trained weights
         tl.files.load_and_assign_npz(name='model_actor.npz', network=self.actor)
         tl.files.load_and_assign_npz(name='model_critic.npz', network=self.critic)
+
 
 class Worker(object):
 
@@ -189,11 +195,11 @@ class Worker(object):
                 # visualize Worker_0 during training
                 if self.name == 'Worker_0' and total_step % 30 == 0:
                     self.env.render()
-                s = s.astype('float32') # double to float
-                a = self.AC.choose_action(s) 
+                s = s.astype('float32')  # double to float
+                a = self.AC.choose_action(s)
                 s_, r, done, _info = self.env.step(a)
-                
-                s_ = s_.astype('float32') # double to float
+
+                s_ = s_.astype('float32')  # double to float
                 # set robot falls reward to -2 instead of -100
                 if r == -100: r = -2
 
@@ -207,7 +213,7 @@ class Worker(object):
                     if done:
                         v_s_ = 0  # terminal
                     else:
-                        v_s_ = self.AC.critic(s_[np.newaxis, :])[0,0] # reduce dim from 2 to 0
+                        v_s_ = self.AC.critic(s_[np.newaxis, :])[0, 0]  # reduce dim from 2 to 0
 
                     buffer_v_target = []
 
@@ -216,7 +222,7 @@ class Worker(object):
                         buffer_v_target.append(v_s_)
 
                     buffer_v_target.reverse()
-                    
+
                     buffer_s, buffer_a, buffer_v_target = (
                         np.vstack(buffer_s), np.vstack(buffer_a), np.vstack(buffer_v_target)
                     )
@@ -265,7 +271,7 @@ if __name__ == "__main__":
         # ============================= TRAINING ===============================
         t0 = time.time()
         with tf.device("/cpu:0"):
-            
+
             OPT_A = tf.optimizers.RMSprop(LR_A, name='RMSPropA')
             OPT_C = tf.optimizers.RMSprop(LR_C, name='RMSPropC')
 
@@ -283,7 +289,7 @@ if __name__ == "__main__":
         for worker in workers:
             # t = threading.Thread(target=worker.work)
             job = lambda: worker.work(GLOBAL_AC)
-            t = threading.Thread(target=job)  
+            t = threading.Thread(target=job)
             t.start()
             worker_threads.append(t)
         COORD.join(worker_threads)
@@ -306,7 +312,7 @@ if __name__ == "__main__":
             rall = 0
             while True:
                 env.render()
-                s = s.astype('float32') # double to float
+                s = s.astype('float32')  # double to float
                 a = GLOBAL_AC.choose_action(s)
                 s, r, d, _ = env.step(a)
                 rall += r
